Chore: extracted different entities
