+++ crates/demo-core/src/lib.rs
@@ -15,6 +15,7 @@ use std::collections::VecDeque;
 #[cfg(all(feature = "events", target_arch = "wasm32"))]
 use std::{cell::RefCell, rc::Rc};
 
+mod event_filters;
 pub mod info_overlay;
 
 pub const MAX_NODE_COUNT: usize = 2500;
@@ -25,179 +26,20 @@ pub const EVENTS_LIMIT: usize = 500;
 const UI_MARGIN: f32 = 10.0;
 
 #[cfg(feature = "events")]
+use crate::event_filters::EventFilters;
+#[cfg(feature = "events")]
 pub use crossbeam::channel::{unbounded, Receiver, Sender};
 #[cfg(feature = "events")]
 pub use egui_graphs::events::Event;
 
-pub mod settings_local {
-    pub struct SettingsInteraction {
-        pub dragging_enabled: bool,
-        pub hover_enabled: bool,
-        pub node_clicking_enabled: bool,
-        pub node_selection_enabled: bool,
-        pub node_selection_multi_enabled: bool,
-        pub edge_clicking_enabled: bool,
-        pub edge_selection_enabled: bool,
-        pub edge_selection_multi_enabled: bool,
-    }
-    impl Default for SettingsInteraction {
-        fn default() -> Self {
-            Self {
-                dragging_enabled: true,
-                hover_enabled: true,
-                node_clicking_enabled: false,
-                node_selection_enabled: false,
-                node_selection_multi_enabled: false,
-                edge_clicking_enabled: false,
-                edge_selection_enabled: false,
-                edge_selection_multi_enabled: false,
-            }
-        }
-    }
-
-    #[derive(Default)]
-    pub struct SettingsStyle {
-        pub labels_always: bool,
-        pub edge_deemphasis: bool,
-    }
-
-    pub struct SettingsNavigation {
-        pub fit_to_screen_enabled: bool,
-        pub zoom_and_pan_enabled: bool,
-        pub zoom_speed: f32,
-        pub fit_to_screen_padding: f32,
-    }
-    impl Default for SettingsNavigation {
-        fn default() -> Self {
-            Self {
-                fit_to_screen_enabled: true,
-                zoom_and_pan_enabled: false,
-                zoom_speed: 0.1,
-                fit_to_screen_padding: 0.1,
-            }
-        }
-    }
-
-    pub struct SettingsGraph {
-        pub count_node: usize,
-        pub count_edge: usize,
-    }
-    impl Default for SettingsGraph {
-        fn default() -> Self {
-            Self {
-                count_node: 25,
-                count_edge: 50,
-            }
-        }
-    }
-}
-pub use settings_local as settings;
+pub mod settings;
 
 fn info_icon(ui: &mut egui::Ui, tip: &str) {
     ui.add_space(4.0);
     ui.small_button("ℹ").on_hover_text(tip);
 }
 
-pub mod drawers {
-    use crate::{MAX_EDGE_COUNT, MAX_NODE_COUNT};
-    use egui::Ui;
-
-    pub struct GraphCountSliders {
-        pub nodes: usize,
-        pub edges: usize,
-    }
-
-    pub fn graph_count_sliders(
-        ui: &mut Ui,
-        mut v: GraphCountSliders,
-        mut on_change: impl FnMut(i32, i32),
-    ) {
-        let mut delta_nodes: i32 = 0;
-        let mut delta_edges: i32 = 0;
-
-        ui.horizontal(|ui| {
-            let start = v.nodes;
-            ui.label("N");
-            ui.add(egui::Slider::new(&mut v.nodes, 0..=MAX_NODE_COUNT));
-            if ui
-                .small_button("-10")
-                .on_hover_text("Remove 10 nodes (M)")
-                .clicked()
-            {
-                v.nodes = v.nodes.saturating_sub(10);
-            }
-            if ui
-                .small_button("-1")
-                .on_hover_text("Remove 1 node (N)")
-                .clicked()
-            {
-                v.nodes = v.nodes.saturating_sub(1);
-            }
-            if ui
-                .small_button("+1")
-                .on_hover_text("Add 1 node (n)")
-                .clicked()
-            {
-                v.nodes = (v.nodes + 1).min(MAX_NODE_COUNT);
-            }
-            if ui
-                .small_button("+10")
-                .on_hover_text("Add 10 nodes (m)")
-                .clicked()
-            {
-                v.nodes = (v.nodes + 10).min(MAX_NODE_COUNT);
-            }
-            delta_nodes = if v.nodes >= start {
-                i32::try_from(v.nodes - start).unwrap()
-            } else {
-                -i32::try_from(start - v.nodes).unwrap()
-            };
-        });
-
-        ui.horizontal(|ui| {
-            let start = v.edges;
-            ui.label("E");
-            ui.add(egui::Slider::new(&mut v.edges, 0..=MAX_EDGE_COUNT));
-            if ui
-                .small_button("-10")
-                .on_hover_text("Remove 10 edges (R)")
-                .clicked()
-            {
-                v.edges = v.edges.saturating_sub(10);
-            }
-            if ui
-                .small_button("-1")
-                .on_hover_text("Remove 1 edge (E)")
-                .clicked()
-            {
-                v.edges = v.edges.saturating_sub(1);
-            }
-            if ui
-                .small_button("+1")
-                .on_hover_text("Add 1 edge (e)")
-                .clicked()
-            {
-                v.edges = (v.edges + 1).min(MAX_EDGE_COUNT);
-            }
-            if ui
-                .small_button("+10")
-                .on_hover_text("Add 10 edges (r)")
-                .clicked()
-            {
-                v.edges = (v.edges + 10).min(MAX_EDGE_COUNT);
-            }
-            delta_edges = if v.edges >= start {
-                i32::try_from(v.edges - start).unwrap()
-            } else {
-                -i32::try_from(start - v.edges).unwrap()
-            };
-        });
-
-        if delta_nodes != 0 || delta_edges != 0 {
-            on_change(delta_nodes, delta_edges);
-        }
-    }
-}
+mod drawers;
 
 pub struct DemoApp {
     pub g: Graph<(), (), Directed, DefaultIx>,
@@ -1758,105 +1600,5 @@ impl DemoApp {
                 }
             }
         });
-    }
-}
-
-#[cfg(feature = "events")]
-#[derive(Clone)]
-pub struct EventFilters {
-    pub pan: bool,
-    pub zoom: bool,
-    pub node_move: bool,
-    pub node_drag_start: bool,
-    pub node_drag_end: bool,
-    pub node_hover_enter: bool,
-    pub node_hover_leave: bool,
-    pub node_select: bool,
-    pub node_deselect: bool,
-    pub node_click: bool,
-    pub node_double_click: bool,
-    pub edge_click: bool,
-    pub edge_select: bool,
-    pub edge_deselect: bool,
-}
-
-#[cfg(feature = "events")]
-impl Default for EventFilters {
-    fn default() -> Self {
-        Self {
-            pan: true,
-            zoom: true,
-            node_move: true,
-            node_drag_start: true,
-            node_drag_end: true,
-            node_hover_enter: true,
-            node_hover_leave: true,
-            node_select: true,
-            node_deselect: true,
-            node_click: true,
-            node_double_click: true,
-            edge_click: true,
-            edge_select: true,
-            edge_deselect: true,
-        }
-    }
-}
-
-#[cfg(feature = "events")]
-impl EventFilters {
-    pub fn enabled_for(&self, e: &Event) -> bool {
-        use Event::*;
-        match e {
-            Pan(_) => self.pan,
-            Zoom(_) => self.zoom,
-            NodeMove(_) => self.node_move,
-            NodeDragStart(_) => self.node_drag_start,
-            NodeDragEnd(_) => self.node_drag_end,
-            NodeHoverEnter(_) => self.node_hover_enter,
-            NodeHoverLeave(_) => self.node_hover_leave,
-            NodeSelect(_) => self.node_select,
-            NodeDeselect(_) => self.node_deselect,
-            NodeClick(_) => self.node_click,
-            NodeDoubleClick(_) => self.node_double_click,
-            EdgeClick(_) => self.edge_click,
-            EdgeSelect(_) => self.edge_select,
-            EdgeDeselect(_) => self.edge_deselect,
-        }
-    }
-    pub fn is_event_str_enabled(&self, ev: &str) -> Option<bool> {
-        if ev.starts_with("Pan") {
-            Some(self.pan)
-        } else if ev.starts_with("Zoom") {
-            Some(self.zoom)
-        } else if ev.starts_with("NodeMove") {
-            Some(self.node_move)
-        } else if ev.starts_with("NodeDragStart") {
-            Some(self.node_drag_start)
-        } else if ev.starts_with("NodeDragEnd") {
-            Some(self.node_drag_end)
-        } else if ev.starts_with("NodeHoverEnter") {
-            Some(self.node_hover_enter)
-        } else if ev.starts_with("NodeHoverLeave") {
-            Some(self.node_hover_leave)
-        } else if ev.starts_with("NodeSelect") {
-            Some(self.node_select)
-        } else if ev.starts_with("NodeDeselect") {
-            Some(self.node_deselect)
-        } else if ev.starts_with("NodeClick") {
-            Some(self.node_click)
-        } else if ev.starts_with("NodeDoubleClick") {
-            Some(self.node_double_click)
-        } else if ev.starts_with("EdgeClick") {
-            Some(self.edge_click)
-        } else if ev.starts_with("EdgeSelect") {
-            Some(self.edge_select)
-        } else if ev.starts_with("EdgeDeselect") {
-            Some(self.edge_deselect)
-        } else {
-            None
-        }
-    }
-    pub fn purge_disabled(&self, events: &mut Vec<String>) {
-        events.retain(|ev| self.is_event_str_enabled(ev.as_str()).unwrap_or(true));
     }
 }
